Refactor evaluations module
